rag_rag/pyproject.toml
[project]
name = "rag-rag"
version = "0.1.0"
description = "Enterprise RAG Pipeline Service with LangGraph"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
authors = [
    {name = "RAG Team"}
]
keywords = ["rag", "langgraph", "llm", "retrieval"]

dependencies = [
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "chromadb>=0.5.0",
    "whoosh>=2.7.4",
    "neo4j>=5.0.0",
    "aiosqlite>=0.20.0",
    "aiofiles>=24.1.0",
    "aiohttp>=3.10.0",
    "pyyaml>=6.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "watchdog>=4.0.0",
    "dashscope>=1.20.0",
    "rank-bm25>=0.2.2",
    "tenacity>=8.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rag_rag"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true